Update DRMV2.0
Works fine but I think there is a problem with the measurements as it is returning a value of 1.66 for verowhite, so something to do with the measurement times in the calc function maybe? As all of the frequency readings are correct. Going to do some tests with the old program to see the differences.